Add __init__ file
This is an app entrypoint, containing asyncio webserver which calls `messages._handle_message` for every received Slack message. It also initialized Hydra client and passes it to messages class.
The app initializes Prometheus counters, which are used by messages class
